integrations/platform/reporting/looker/export_job/requirements.txt
requests==2.32.3
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1

